fix: menu corrections
