+++ ChatBot/Main.js
@@ -7,6 +7,8 @@ const fetchProduct = require("@mod/ask_busqueda");
 const pushProduct = require("@mod/ask_push")
 require('dotenv').config();
 let boton=require('@mod/buttons');
+
+
 const bot = new TeleBot({
     token:process.env.TOKEN,
     usePlugins: ['commandButton', 'askUser']
@@ -33,11 +35,6 @@ const bot = new TeleBot({
         bot.sendMessage(msg.from.id,`${msg.from.first_name}, actualmente aceptamos`,{replyMarkup});
     });
 
-    bot.on('/zona',msg=>{
-        let replyMarkup=boton(msg.data,bot);
-        bot.sendMessage(msg.from.id, 'Trabajamos de 9:00 A.M - 6:00 P.M de lunes a viernes, y realizamos entregas en:', {replyMarkup});
-    });
-
     bot.on('/view',msg=>{
         getTrolley(msg,bot);
     });
@@ -53,14 +50,29 @@ const bot = new TeleBot({
 
     bot.on('/search',msg=>{
         const id=msg.from.id;
-        bot.sendMessage(msg.from.id,"Ingrese un ID para buscar producto", {ask:'busqueda'});
+        bot.sendMessage(msg.from.id,"Debes escribir el 🆔 para mostrar un producto", {ask:'busqueda'});
     });
 
     bot.on('ask.busqueda', msg=>{
         fetchProduct(msg,bot);
     });
 
+
     //-------------------------------------- Enviar mapas
+
+
+    bot.on('/zona',msg=>{
+        let replyMarkup=boton(msg.data,bot);
+        bot.sendMessage(msg.from.id, 'También tenemos delivery gratis en estas ubicaciones 🔥, selecciona una para saber como llegar');
+        bot.sendMessage(msg.from.id, '¡Trabajamos de 9:00 A.M a 6:00 P.M de lunes a viernes! 😎', {replyMarkup})
+    });
+
+    
+
+    bot.on('/otrasZona',msg=>{
+        let replyMarkup=boton(msg.data,bot);
+        bot.sendMessage(msg.from.id, 'Aprende a llegar nuestros puntos de entrega', {replyMarkup});
+    });
 
     bot.on('/map1', msg => {
         let replyMarkup =boton(msg.data, bot);
@@ -76,60 +88,4 @@ const bot = new TeleBot({
         bot.sendLocation(msg.from.id, [10.59882925256173, -71.65087523348457], {replyMarkup});
     });
 
-    //-------------------------------------- Generar factura
-
-    // bot.on('/fact',msg=>{
-    //     let replyMarkup = bot.inlineKeyboard(boton(msg.data,bot))
-    //     bot.sendMessage(msg.from.id,"Reviza tu factura para confirmar tu pago", {replyMarkup});
-    // });
-
-    bot.on('/fact', (msg) => {
-        const inlineKeyboard = boton(msg.data,bot);
-        return bot.sendInvoice(msg.from.id, {
-            title: 'Factura de compra',
-            description: 'Node API Store',
-            payload: 'telebot-test-invoice',
-            providerToken: '284685063:TEST:MGZlNzhkZjkwOGZm',
-            startParameter: 'pay',
-            currency: 'USD',
-            prices: [
-                {label: 'Tea', amount: 125},
-                {label: 'For testing!', amount: 1250},
-            ],
-            need: {email: true, phoneNumbe: false},
-            // needShippingAdress: true,
-            // isFlexible: false,
-            // sendEmailToProvider: true,
-            replyMarkup: inlineKeyboard
-        }).then(data => {
-            console.log('OK', data);
-        }).catch(error => {
-            console.log('ERROR', error);
-        });
-    
-    });
-    
-    // bot.on('shippingQuery', (msg) => {
-    //     console.log('shippingQuery', msg);
-    // });
-    
-    // bot.on('preShippingQuery', (msg) => {
-    //     console.log('preShippingQuery', msg);
-    
-    //     const id = msg.id;
-    //     const isOk = true;
-    
-    //     return bot.answerPreCheckoutQuery(id, isOk);
-    
-    // });
-    
-    bot.on('successfulPayment', (msg) => {
-        console.log('successfulPayment', msg);
-    
-        return bot.sendMessage(msg.from.id, `Gracias por tu compra, ${msg.from.first_name}!`);
-    
-    });
-    
-    
-    
 bot.start();
